Use gfycat-sdk npm module.
