+++ package.json
@@ -5,6 +5,7 @@
     "deep-equal": "1.0.1",
     "eslint-plugin-header": "1.2.0",
     "form-data": "2.3.2",
+    "gfycat-sdk": "1.4.18",
     "harmony-reflect": "1.6.0",
     "isomorphic-fetch": "2.2.1",
     "mime-db": "1.33.0",
